restructure code package
