--- app/src/main/java/com/example/pengalatdite/aplikasisederhana/ItemViewMvc.java
@@ -1,9 +1,0 @@
-package com.example.pengalatdite.aplikasisederhana;
-
-import android.view.View;
-
-public interface ItemViewMvc extends ViewMvc {
-    View getRootView();
-
-    void bindData(DataNews dataNews, String position);
-}
