+++ app/src/main/java/com/example/pengalatdite/aplikasisederhana/screen/newslist/newslistitem/ItemViewMvc.java
@@ -1,0 +1,12 @@
+package com.example.pengalatdite.aplikasisederhana.screen.newslist.newslistitem;
+
+import android.view.View;
+
+import com.example.pengalatdite.aplikasisederhana.data.DataNews;
+import com.example.pengalatdite.aplikasisederhana.screen.common.ViewMvc;
+
+public interface ItemViewMvc extends ViewMvc {
+    View getRootView();
+
+    void bindData(DataNews dataNews, String position);
+}
